mpHGreX (nie zachowuje 0 na poczatku), permutations liczy ale wolno przy zuzyciu 3gb ramu na watek

+++ Optimization/WarehouseMode/Warehouse.cs
@@ -25,7 +25,7 @@ namespace OptimizationMethods.WarehouseMode
                 (population, matrix) =>
                 {
                     double[] fitness = new double[population.Length];
-                    Parallel.For((long) 0, population.Length, i =>
+                    Parallel.For( 0, population.Length, i =>
                     {
                         fitness[i] = Fitness.CalculateAllOrdersFitness(orders, population[i], distancesMatrix, warehouseParameters.FitnessGeneticAlgorithmParameters);
                     });
